requirements.txt
flask
requests
gunicorn
google-generativeai
python-dotenv
